Rename matching to quizz_matching.
