+++ modules/question_types/quizz_matching/quizz_matching.module
@@ -6,34 +6,19 @@ use Drupal\quizz_question\Entity\QuestionType;
  * @file
  * Matching question type for quiz module
  *
- * Allows the creation of matching questions, which associate one term
- * with another
+ * Allows the creation of matching questions, which associate one term with another.
  */
-
-/**
- * Implements hook_help().
- */
-function matching_help($path, $args) {
-  switch ($path) {
-    case 'admin/modules#description':
-      return t('Matching question type for quiz module.');
-    case 'admin/help#matching':
-      return t('A question type for the quiz module: allows you to create matching type questions, which connect terms with one another.');
-    default:
-      break;
-  }
-}
 
 /**
  * Implements hook_quiz_question_info().
  */
-function matching_quiz_question_info() {
+function quizz_matching_quiz_question_info() {
   return array(
       'matching' => array(
           'name'              => t('Matching'),
           'description'       => t('Matching question type for quiz module. A question type for the quiz module: allows you to create matching type questions, which connect terms with one another.'),
-          'question provider' => 'MatchingQuestion',
-          'response provider' => 'MatchingResponse',
+          'question provider' => 'Drupal\quizz_matching\MatchingQuestion',
+          'response provider' => 'Drupal\quizz_matching\MatchingResponse',
           'module'            => 'quizz_question', // All wrapper functions are in that module.
       ),
   );
@@ -44,8 +29,8 @@ function matching_quiz_question_info() {
  *
  * @return FAPI array
  */
-function matching_quiz_question_config(QuestionType $question_type) {
-  $form = array('#validate' => array('matching_config_validate'));
+function quizz_matching_quiz_question_config(QuestionType $question_type) {
+  $form = array('#validate' => array('quizz_matching_config_validate'));
 
   $form['quiz_matching_form_size'] = array(
       '#type'          => 'textfield',
@@ -67,7 +52,7 @@ function matching_quiz_question_config(QuestionType $question_type) {
 /**
  * Validate the matching config form values
  */
-function matching_config_validate($form, $form_state) {
+function quizz_matching_config_validate($form, $form_state) {
   if (!quizz_valid_integer($form_state['values']['quiz_matching_form_size'], 2, 50)) {
     form_set_error('quiz_matching_form_size', t('The number of questions must be between 2 and 50'));
   }
@@ -76,53 +61,32 @@ function matching_config_validate($form, $form_state) {
 /**
  * Implements hook_theme().
  */
-function matching_theme() {
+function quizz_matching_theme() {
   return array(
       'matching_question_form'    => array(
           'render element' => 'form',
-          'path'           => drupal_get_path('module', 'matching') . '/theme',
-          'file'           => 'matching.theme.inc',
+          'path'           => drupal_get_path('module', 'quizz_matching') . '/theme',
+          'file'           => 'quizz_matching.theme.inc',
       ),
       'matching_response'         => array(
           'variables' => array('metadata' => NULL, 'data' => NULL),
-          'path'      => drupal_get_path('module', 'matching') . '/theme',
-          'file'      => 'matching.theme.inc',
+          'path'      => drupal_get_path('module', 'quizz_matching') . '/theme',
+          'file'      => 'quizz_matching.theme.inc',
       ),
       'matching_subquestion_form' => array(
           'render element' => 'form',
-          'path'           => drupal_get_path('module', 'matching') . '/theme',
-          'file'           => 'matching.theme.inc',
+          'path'           => drupal_get_path('module', 'quizz_matching') . '/theme',
+          'file'           => 'quizz_matching.theme.inc',
       ),
       'matching_match_node_view'  => array(
           'variables' => array('subquestions' => NULL),
-          'path'      => drupal_get_path('module', 'matching') . '/theme',
-          'file'      => 'matching.theme.inc',
+          'path'      => drupal_get_path('module', 'quizz_matching') . '/theme',
+          'file'      => 'quizz_matching.theme.inc',
       ),
       'matching_answering_form'   => array(
           'render element' => 'form',
-          'path'           => drupal_get_path('module', 'matching') . '/theme',
+          'path'           => drupal_get_path('module', 'quizz_matching') . '/theme',
           'template'       => 'matching-answering-form',
       ),
   );
-}
-
-/**
- * Implements hook_field_extra_fields().
- */
-function matching_field_extra_fields() {
-  $extra['node']['matching'] = array(
-      'form' => array(
-          'settings' => array(
-              'label'       => t('Settings'),
-              'description' => t('The settings for the matching question.'),
-              'weight'      => -5,
-          ),
-          'match'    => array(
-              'label'       => t('Answer'),
-              'description' => t('The sets of matches for this question.'),
-              'weight'      => -4,
-          ),
-      ),
-  );
-  return $extra;
 }
