+++ modules/question_types/quizz_matching/theme/matching-answering-form.tpl.php
