+++ modules/question_types/quizz_matching/theme/quizz_matching.theme.inc
@@ -1,30 +1,23 @@
 <?php
 
-
-/**
- * @file
- *   Themes for the matching module.
- */
-
 /**
  * Theme the matching question form.
  *
  * @param $form
- *   drupal form array
  */
 function theme_matching_question_form($variables) {
   $form = $variables['form'];
   $rows = array();
   $header = array(
-    'question' => t('Question'),
-    'answer' => t('Correct answer'),
-    'feedback' => t('Feedback'),
+      'question' => t('Question'),
+      'answer'   => t('Correct answer'),
+      'feedback' => t('Feedback'),
   );
   foreach (element_children($form) as $key) {
     $rows[] = array(
-      'question' => drupal_render($form[$key]['question']),
-      'answer' => drupal_render($form[$key]['answer']),
-      'feedback' => drupal_render($form[$key]['feedback']),
+        'question' => drupal_render($form[$key]['question']),
+        'answer'   => drupal_render($form[$key]['answer']),
+        'feedback' => drupal_render($form[$key]['feedback']),
     );
   }
   // Theme output and display to screen.
@@ -90,8 +83,8 @@ function theme_matching_response($variables) {
       }
       // TODO Please change this theme call to use an associative array for the $variables parameter.
       $data[$id]['is_correct'] = array(
-        'data' => theme($theme),
-        'class' => 'quiz-summary-qcell',
+          'data'  => theme($theme),
+          'class' => 'quiz-summary-qcell',
       );
     }
   }
